requirements.txt
asgiref==3.7.2
bleach==6.1.0
defusedxml==0.7.1
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
webencodings==0.5.1
